requirements.txt
requests-cache==0.5.2
tox==4.0.0a9
pyyaml
bs4